pyproject.toml
[project]
name = "exam-system"
version = "1.0.0"
description = "JEE Mock CBT Portal"
dependencies = [
    "fastapi",
    "uvicorn",
    "firebase-admin",
    "PyJWT",
    "bcrypt",
    "python-dotenv",
]

[tool.setuptools]
py-modules = ["main"]

# Equivalent to your "scripts": {"start": "node server.js"}
[project.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"